# Add sample_single to Range