--- a/src/distributions/uniform.rs
+++ b/src/distributions/uniform.rs
@@ -54,12 +54,17 @@ pub fn codepoint<R: Rng+?Sized>(rng: &mut R) -> char {
 /// TODO: should this be removed in favour of `AsciiWordChar`?
 #[inline]
 pub fn ascii_word_char<R: Rng+?Sized>(rng: &mut R) -> char {
-    use sequences::Choose;
+    const RANGE: u32 = 26 + 26 + 10;
     const GEN_ASCII_STR_CHARSET: &'static [u8] =
         b"ABCDEFGHIJKLMNOPQRSTUVWXYZ\
             abcdefghijklmnopqrstuvwxyz\
             0123456789";
-    *GEN_ASCII_STR_CHARSET.choose(rng).unwrap() as char
+    loop {
+        let var = rng.next_u32() & 0x3F;
+        if var < RANGE {
+            return GEN_ASCII_STR_CHARSET[var as usize] as char
+        }
+    }
 }