@pitdicker 2017-12-12
Just added a little optimisation to `ascii_word_char`. It's range is close enough to a power of two that a bitmask is more efficient.

With `choose` I see a optimisation, but don't know if it is possible. It needs an `usize` in some range to use as slice index. on x86_64 this is the same as an `u64`, which is twice as slow as `u32`. It would help if we could assume the slice length < 2^32. Would it be ok to document: "`Choose` will only pick from the first 2^32 values of a slice"?