@dhardy 2017-12-12
Hmm, not sure how to answer this. I'm not so happy with the use of `usize` for container sizes everywhere since for the *vast* majority of uses `u32` is sufficient, but it's what we have, and it seems odd doing something different here.