@pitdicker 2017-12-12
With `i128_support` on x86_64 a range reduction on `u64` should be about as fast as on `u32`.

For cryptographic RNG's generating `u64`'s instead of `u32`'s should be about 50~60 percent slower. It is not 100% because of (thanks to) the overhead of reading from the buffered the results. At least with an improvement I made for HC-128, that should also work for ISAAC and ChaCha.

Small fast RNG's mostly need to operate on 64-bit words for good statistical quality. There are some really fast ones where on x86_64 generating `u64`'s has the same excellent performance as generating `u32`'s.

So on second thought using `usize` on x86_64 should eventually have similar performance as using an `u32` as index. We should just leave `choose` as it is, without subtleties.