@pitdicker 2017-12-12
Now I wonder if the optimization of `ascii_word_char` is worth it. It is only about 10% faster then when we have better RNG's and `i128_support`.

I do like that it removes a complex dependency between `Uniform`, `sequences`, `Sample` and `Range`. Before the chain looked like: `Uniform::ascii_word_char` → `sequences::Choose` → `Sample::gen_range` → `Range::sample::single`. I think it makes some sense to end up with distributions that only depend on `rand_core` (and `std`/`core`).