@dhardy 2017-12-12
Yes, the optimisation/standalone impl for `ascii_word_char` is a good idea in my opinion.

Not sure what you mean here?

> Also I tried to make sample_single have a default implementation, like in RangeFloat, but couldn't get it to work.