@dhardy 2017-12-12
Ah. It's actually quite simple:
```
error[E0277]: the trait bound `Self: core::marker::Sized` is not satisfied                                                             
   --> src/distributions/range.rs:107:13                                                                                               
    |                                                                                                                                  
107 |         let range: Self = RangeImpl::new(low, high);                                                                             
    |             ^^^^^ `Self` does not have a constant size known at compile-time
```
What the compiler is saying is that you can only construct *sized types*; here you are trying to construct `range` of type `RangeImpl`, but it doesn't know that the `RangeImpl` trait can only be implemented for `Sized` types (e.g. you could try to `impl RangeImpl for [u8]`). You can fix this by specifying `trait RangeImpl: Sized { ... }`; this restricts implementations to sized types.