@pitdicker 2017-12-12
I can't really imagine a `RangeImpl` for slices, so a `Sized` bound should not really be a restriction. Thank you, 20 more errors and I will start to get a feeling for traits :smile:. Will add a commit tomorrow.