@pitdicker 2017-12-13
Now that you mention it, I did not add any tests... The tests for `choose` cover testing if some values give plausible results, and if the values are in range.