@dhardy 2017-12-14
No, properly testing distributions would appear to be much more difficult than writing the distribution code.